# docs/swagger.yaml
openapi: 3.0.3
info:
  title: Smriti API
  description: Smarter Home for all your Photos and Videos
  version: 24.06.24
servers:
  - url: https://localhost:5001
...
paths:
...
  /v1/sharing/{id}:
    get:
      tags:
        - Sharing
      parameters:
        - name: id
          in: path
          description: ID of shared album
          required: true
          schema:
            type: string
            format: uuid
      summary: Get an existing shared album
      description: Get an existing shared album
      operationId: getSharedAlbum
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Album'
        '400':
          description: Bad Request
        '401':
          description: Unauthenticated
        '404':
          description: Not Found
        '500':
          description: Internal Server Error
  /v1/sharing/{id}/mediaItems:
    get:
      tags:
        - Sharing
      parameters:
        - name: id
          in: path
          description: ID of shared album
          required: true
          schema:
            type: string
            format: uuid
      summary: Get mediaItems of an existing shared album
      description: Get mediaItems of an existing shared album
      operationId: getSharedAlbumMediaItems
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MediaItems'
        '400':
          description: Bad Request
        '401':
          description: Unauthenticated
        '404':
          description: Not Found
        '500':
          description: Internal Server Error
...
  /v1/users:
    get:
      tags:
        - User Management
      summary: Get all existing users
      description: Get all existing users
      operationId: getUsers
      security:
        - basicAuth: []
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Users'
        '400':
          description: Bad Request
        '401':
          description: Unauthenticated
        '500':
          description: Internal Server Error
    post:
      tags:
        - User Management
      summary: Create a new user
      description: Create a new user
      operationId: createUser
      security:
        - basicAuth: []
      requestBody:
        description: User request body
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRequest'
      responses:
        '201':
          description: Created
        '400':
          description: Bad Request
        '401':
          description: Unauthenticated
        '500':
          description: Internal Server Error
  /v1/users/{id}:
    get:
      tags:
        - User Management
      parameters:
        - name: id
          in: path
          description: ID of user
          required: true
          schema:
            type: string
            format: uuid
      summary: Get an existing user
      description: Get an existing user
      operationId: getUser
      security:
        - basicAuth: []
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Bad Request
        '401':
          description: Unauthenticated
        '404':
          description: Not Found
        '500':
          description: Internal Server Error
    put:
      tags:
        - User Management
      parameters:
        - name: id
          in: path
          description: ID of user
          required: true
          schema:
            type: string
            format: uuid
      summary: Update an existing user
      description: Update an existing user
      operationId: updateUser
      security:
        - basicAuth: []
      requestBody:
        description: User request body
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRequest'
      responses:
        '204':
          description: No Content
        '400':
          description: Bad Request
        '401':
          description: Unauthenticated
        '404':
          description: Not Found
        '500':
          description: Internal Server Error
    delete:
      tags:
        - User Management
      parameters:
        - name: id
          in: path
          description: ID of user
          required: true
          schema:
            type: string
            format: uuid
      summary: Delete an existing user
      description: Delete an existing user
      operationId: deleteUser
      security:
        - basicAuth: []
      responses:
        '204':
          description: No Content
        '400':
          description: Bad Request
        '401':
          description: Unauthenticated
        '404':
          description: Not Found
        '500':
          description: Internal Server Error
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    basicAuth:
      type: http
      scheme: basic
  schemas:
...
    MediaItem:
      type: object
      properties:
        id:
          readOnly: true
          type: string
          format: uuid
          example: 102dc522de-1669-48a4-8270-4950a0237599
        userId:
          readOnly: true
          type: string
          format: uuid
          example: 102dc522de-1669-48a4-8270-4950a0237599
        filename:
          readOnly: true
          type: string
          example: IMG_4942.jpg
        description:
          readOnly: true
          type: string
          example: Picture of a pizza from New York
        mimeType:
          readOnly: true
          type: string
          example: image/jpeg
        placeholder:
          readOnly: true
          type: string
          example: base-64-placeholder-image
        sourceUrl:
          readOnly: true
          type: string
          example: https://cdn.smriti/source
        previewUrl:
          readOnly: true
          type: string
          example: https://cdn.smriti/preview
        thumbnailUrl:
          readOnly: true
          type: string
          example: https://cdn.smriti/thumbnail
        favourite:
          type: boolean
          example: false
        shared:
          type: boolean
          example: true
        hidden:
          type: boolean
          example: false
        deleted:
          type: boolean
          example: false
        status:
          readOnly: true
          type: string
          enum: ["UNSPECIFIED", "PROCESSING", "READY", "FAILED"]
        mediaItemType:
          readOnly: true
          type: string
          enum: ["photo", "video"]
        mediaItemCategory:
          readOnly: true
          type: string
          enum: ["default", "screenshot", "panorama", "slow", "motion", "live"]
        width:
          readOnly: true
          type: integer
          example: 1080
        height:
          readOnly: true
          type: integer
          example: 720
        creationTime:
          readOnly: true
          type: string
          format: date-time
          example: "2022-09-14T16:58:28Z"
        latitude:
          readOnly: true
          type: number
          example: 19.384982
        longitude:
          readOnly: true
          type: number
          example: -70.638324
        cameraMake:
          readOnly: true
          type: string
          example: Apple
        cameraModel:
          readOnly: true
          type: string
          example: iPhone 12
        focalLength:
          readOnly: true
          type: string
          example: 5.49 mm
        apertureFNumber:
          readOnly: true
          type: string
          example: f/1.8
        isoEquivalent:
          readOnly: true
          type: string
          example: ISO758
        exposureTime:
          readOnly: true
          type: string
          example: 1/20
        fps:
          readOnly: true
          type: string
          example: 60
        createdAt:
          readOnly: true
          type: string
          format: date-time
          example: "2022-09-14T16:58:28Z"
        updatedAt:
          readOnly: true
          type: string
          format: date-time
          example: "2022-09-14T16:58:28Z"
...
    Album:
      type: object
      properties:
        id:
          readOnly: true
          type: string
          format: uuid
          example: 102dc522de-1669-48a4-8270-4950a0237599
        userId:
          readOnly: true
          type: string
          format: uuid
          example: 102dc522de-1669-48a4-8270-4950a0237599
        name:
          type: string
          example: Jane's Wedding
        description:
          type: string
          example: Pictures from Wedding of Jane and Doe
        coverMediaItemId:
          type: string
          format: uuid
          example: 102dc522de-1669-48a4-8270-4950a0237599
        shared:
          type: boolean
          example: true
        hidden:
          type: boolean
          example: false
        mediaItemsCount:
          type: integer
          example: 42
        createdAt:
          readOnly: true
          type: string
          format: date-time
          example: "2022-09-14T16:58:28Z"
        updatedAt:
          readOnly: true
          type: string
          format: date-time
          example: "2022-09-14T16:58:28Z"
        coverMediaItem:
          $ref: '#/components/schemas/MediaItem'
...
    UserRequest:
      type: object
      properties:
        name:
          type: string
          example: John Doe
        username:
          type: string
          example: johndoe
        password:
          type: string
          example: somesecrethash
    User:
      type: object
      properties:
        id:
          readOnly: true
          type: string
          format: uuid
          example: 102dc522de-1669-48a4-8270-4950a0237599
        name:
          type: string
          example: John Doe
        username:
          type: string
          example: johndoe
        password:
          type: string
          example: somesecrethash
        createdAt:
          readOnly: true
          type: string
          format: date-time
          example: "2022-09-14T16:58:28Z"
        updatedAt:
          readOnly: true
          type: string
          format: date-time
          example: "2022-09-14T16:58:28Z"
...
    MediaItems:
      type: array
      items:
        $ref: '#/components/schemas/MediaItem'
...
    Users:
      type: array
      items:
        $ref: '#/components/schemas/User'
...